+++ examples/sos-sso/sql/4-privileges.sql
@@ -1,19 +1,17 @@
-GRANT USAGE ON `sos-sso-production`.* TO 'nss-maria-user'@'localhost';
-GRANT USAGE ON `sos-sso-production`.* TO 'nss-maria-root'@'localhost';
-
--- GRANT SELECT (col1), INSERT (col1,col2) ON mydb.mytbl TO 'someuser'@'somehost';
-
-GRANT SELECT (`username`, `id`, `gid`, `gecos`, `homedir`, `shell`)
+GRANT SELECT (`active`, `username`, `id`, `gid`, `gecos`, `homedir`, `shell`)
   ON `sos-sso-production`.`users` TO 'nss-maria-user'@'localhost';
+GRANT SELECT (`active`, `username`, `id`, `gid`, `gecos`, `homedir`, `shell`)
+  ON `sos-sso-production`.`users` TO 'nss-maria-user'@'172.%';
 
-GRANT SELECT (`id`, `name`, `password`)
-  ON `sos-sso-production`.`groups` TO 'nss-maria-user'@'localhost';
+GRANT SELECT (`id`, `name`, `password`) ON `sos-sso-production`.`groups` TO 'nss-maria-user'@'localhost';
+GRANT SELECT (`id`, `name`, `password`) ON `sos-sso-production`.`groups` TO 'nss-maria-user'@'172.%';
 
--- root only
+GRANT SELECT (`group_id`, `user_id`) ON `sos-sso-production`.`groups_users` TO 'nss-maria-user'@'localhost';
+GRANT SELECT (`group_id`, `user_id`) ON `sos-sso-production`.`groups_users` TO 'nss-maria-user'@'172.%';
+
 GRANT SELECT (`username`,`password`,`lstchg`,`min`,`max`,`warn`,`inact`,`expire`,`flag`)
   ON `sos-sso-production`.`users` TO 'nss-maria-root'@'localhost';
-
-
-GRANT ALL PRIVILEGES ON `sos-sso-production`.* TO 'nss-maria-user'@'localhost';
+GRANT SELECT (`username`,`password`,`lstchg`,`min`,`max`,`warn`,`inact`,`expire`,`flag`)
+  ON `sos-sso-production`.`users` TO 'nss-maria-root'@'172.%';
 
 FLUSH PRIVILEGES;