+++ scripts/compile_and_test.sh
@@ -11,9 +11,11 @@ fi
 
 SUDO_COMMAND=$([[ -z $TEST_USE_SUDO || $TEST_USE_SUDO = "0" ]] && echo "" || echo "sudo")
 CONFIG_FILE=${CONFIG_FILE:-libnss-maria.conf}
+CONFIG_ROOT_FILE=${CONFIG_ROOT_FILE:-libnss-maria-root.conf}
 
 echo "COMPILE_ONLY: $COMPILE_ONLY"
 echo "CONFIG_FILE: $CONFIG_FILE"
+echo "CONFIG_ROOT_FILE: $CONFIG_ROOT_FILE"
 echo "DC_FILE: $DC_FILE"
 echo "DOCKER: $DOCKER"
 echo "PREDEFINED_DB: $PREDEFINED_DB"
@@ -25,7 +27,7 @@ mkdir "$HOME_PATH/$TARGET"
 cd ${HOME_PATH}/${TARGET} && cmake -D CMAKE_BUILD_TYPE=${TARGET} .. && make && ctest --verbose
 ORIG_RETURN_CODE=$?
 
-if [[ $? -eq 0 && -z $COMPILE_ONLY ]]; then
+if [[ $ORIG_RETURN_CODE -eq 0 && -z $COMPILE_ONLY ]]; then
   $SUDO_COMMAND rm /lib/libnss_maria.so*
   $SUDO_COMMAND cp -bf /home/libnss-maria/${TARGET}/src/libnss_maria.so.2.0.0 /lib/libnss_maria.so.2.0.0 &&
   $SUDO_COMMAND ln -fs /lib/libnss_maria.so.2.0.0 /lib/libnss_maria.so &&
@@ -41,7 +43,9 @@ if [[ $? -eq 0 && -z $COMPILE_ONLY ]]; then
 
   $SUDO_COMMAND cp -bf /home/libnss-maria/examples/sos-sso/nsswitch.conf /etc &&
   $SUDO_COMMAND cp -bf /home/libnss-maria/examples/sos-sso/$CONFIG_FILE /etc/libnss-maria.conf &&
-  $SUDO_COMMAND cp -bf /home/libnss-maria/examples/sos-sso/libnss-maria-root.conf /etc
+  $SUDO_COMMAND cp -bf /home/libnss-maria/examples/sos-sso/$CONFIG_ROOT_FILE /etc/libnss-maria-root.conf
+  $SUDO_COMMAND chmod 644 /etc/nsswitch.conf /etc/libnss-maria.conf
+  $SUDO_COMMAND chmod 600 /etc/libnss-maria-root.conf
 
   ${HOME_PATH}/test/integration/nss_module.bats
 else