+++ src/configuration/maria_config.h
@@ -3,6 +3,6 @@
 #include "maria_config_struct.h"
 #include "maria_config_helpers.h"
 
-int maria_read_config_file(Maria_config* config, const char *path);
+int maria_read_config(Maria_config* config, const char *path, const char *root_path);
 
 #endif