+++ src/configuration/maria_config_helpers.h
@@ -12,5 +12,6 @@ void maria_initialize_config(Maria_config *config);
 void maria_load_string_setting(config_t libconfig_object, char *destination, const char *selector);
 void maria_load_int64_setting(config_t libconfig_object, long long *destination, const char *selector);
 int maria_set_config_from_file(const char *path, Maria_config *config);
+int maria_set_root_config_from_file(const char *path, Maria_config *config);
 
 #endif